pyproject.toml
[tool.poetry]
name = "goingfast"
version = "0.1.0"
description = ""
authors = ["Batista Harahap <batista@bango29.com>"]
license = "GPL"

[tool.poetry.dependencies]
python = "^3.10"
aiodns = "3.0.0"
aiofiles = "0.8.0"
aiohttp = "3.8.1"
aiosignal = "1.2.0"
anyio = "3.6.1"
async-timeout = "4.0.2"
attrs = "22.1.0"
ccxt = "1.91.96"
certifi = "2022.6.15"
cffi = "1.15.1"
chardet = "5.0.0"
charset-normalizer = "2.1.0"
cryptography = "37.0.4"
frozenlist = "1.3.1"
h11 = "0.12.0"
h2 = "4.1.0"
hpack = "4.0.0"
hstspreload = "2022.8.1"
httpcore = "0.15.0"
httptools = "0.4.0"
httpx = "0.23.0"
hyperframe = "6.0.1"
idna = "3.3"
multidict = "6.0.2"
pycares = "4.2.1"
pycparser = "2.21"
requests = "2.28.1"
rfc3986 = "1.5.0"
sanic = "22.6.1"
sanic-routing = "22.3.0"
six = "1.16.0"
sniffio = "1.2.0"
telepot = "12.7"
ujson = "5.4.0"
urllib3 = "1.26.11"
uvloop = "0.16.0"
websockets = "10.3"
yarl = "1.7.2"
pyfunctional = "^1.4.3"
python-binance = "^1.0.16"
numpy = "^1.23.1"
TA-Lib = "^0.4.24"
pyfiglet = "^0.8.post1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[tool.black]
line-length = 120
target-version = ["py310"]
skip-string-normalization = true
skip-magic-trailing-comma = true
exclude = '''
    /(
        \.git
      | \.hg
      | \.mypy_cache
      | \.tox
      | \.venv
      | _build
      | buck-out
      | build
      | dist
      # The following are specific to Black, you probably don't want those.
      | blib2to3
      | tests/data
    )/
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
